Make the HostID include the RPC port
This change allows the user to run multiple agents on a single host by
setting the RPC listen port (via the --listen argument).

This change was made in order to support testing a very large number of
agents without requiring a commensurate number of hosts.

Cherry pick changes to allow for multiple agents from MultiAgentBuild